requirements.txt
streamlit
pillow
numpy
cryptography